Moved stata_lsp to module in lua
